fix(chat+nav+context): full-text default, all pages in sidebar, test event loop fix
Context & Sources:
- chat.py: default context_size 'short'→'long' for API calls without context_config
- sources.py: raise hard cap le=100→1000, default 50→1000
- use-sources.ts: page size 30→1000 (all sources in one request)
- notebooks/[id]/page.tsx: default source mode 'full', remove auto-downgrade to 'insights'
- context_builder.py: restore truncate_to_fit with warning log; build_context() helper
  sends 'full content' for sources instead of 'insights'

Navigation — 8 orphaned pages now in sidebar:
- AppSidebar.tsx: Notebooks, Pipeline & CRM, Customers, Contacts,
  Transformations, Research Memory, Publications & Calendar, Advanced Tools

Tests:
- test_canvas_quiz_linkage.py: fix async/sync event loop conflict — TestClient
  mixed with await in @pytest.mark.asyncio caused teardown failures in full suite.
  Changed to sync def + asyncio.run() for repo cleanup calls.

Gitignore: add 4 tool-generated dashboard plan docs

+++ api/routers/chat.py
@@ -494,11 +494,13 @@ async def build_context(request: BuildContextRequest):
                     logger.warning(f"Error processing note {note_id}: {str(e)}")
                     continue
         else:
-            # Default behavior - include all sources and notes with short context
+            # Default: full document text for all sources and notes.
+            # 'short' (title + insights only) was the old default and caused the LLM
+            # to answer questions without seeing document content.
             sources = await notebook.get_sources()
             for source in sources:
                 try:
-                    source_context = await source.get_context(context_size="short")
+                    source_context = await source.get_context(context_size="long")
                     context_data["sources"].append(source_context)
                     total_content += str(source_context)
                 except Exception as e:
@@ -508,7 +510,7 @@ async def build_context(request: BuildContextRequest):
             notes = await notebook.get_notes()
             for note in notes:
                 try:
-                    note_context = note.get_context(context_size="short")
+                    note_context = note.get_context(context_size="long")
                     context_data["notes"].append(note_context)
                     total_content += str(note_context)
                 except Exception as e:
